Implementado sistema de deguelles

+++ src/main/java/cava/model/repository/MaterialRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import cava.model.entity.Jaula;
 import cava.model.entity.Material;
 
-public interface MaterialRepository extends JpaRepository<Material, Integer>{
+public interface MaterialRepository extends JpaRepository<Material, Long>{
 
 }
