+++ src/main/java/cava/model/service/MaterialService.java
@@ -2,6 +2,6 @@ package cava.model.service;
 
 import cava.model.entity.Material;
 
-public interface MaterialService extends InterfaceGenericoCrud<Material, Integer>{
+public interface MaterialService extends InterfaceGenericoCrud<Material, Long>{
 
 }
